+++ packages/beta/src/__tests__/api/monitoringTasksApi.int.spec.ts
@@ -17,6 +17,7 @@ describe('monitoring tasks api', () => {
   const ts = Date.now();
   const monitoringTaskExternalId = `test_mt_${ts}`;
   const monitoringTaskName = `test_mt_${ts}`;
+  const monitoringTaskNameUpdated = monitoringTaskName + '_updated';
   const channelExternalId = `test_channel_mt_${ts}`;
   const sessionsApi = `/api/v1/projects/${process.env.COGNITE_PROJECT}/sessions`;
   const testMtModel: MonitoringTaskDoubleThresholdModelCreate = {
@@ -93,6 +94,34 @@ describe('monitoring tasks api', () => {
     done();
   }, 10000);
 
+  test('upsert monitoring task', async () => {
+    const sessionsRes = await client.post<SessionsResponse>(sessionsApi, {
+      data: {
+        items: [
+          {
+            clientId: process.env.COGNITE_CLIENT_ID,
+            clientSecret: process.env.COGNITE_CLIENT_SECRET,
+          },
+        ],
+      },
+    });
+
+    const response = await client.monitoringTasks.upsert([
+      {
+        externalId: monitoringTaskExternalId,
+        name: monitoringTaskNameUpdated,
+        channelId: channel.id,
+        interval: testMtInterval,
+        nonce: sessionsRes?.data?.items[0]?.nonce,
+        overlap: testMtOverlap,
+        model: testMtModel,
+      },
+    ]);
+
+    expect(response.length).toBe(1);
+    expect(response[0].name).toBe(monitoringTaskNameUpdated);
+  });
+
   test('list all monitoring tasks', async () => {
     const response = await client.monitoringTasks.list({
       filter: {},
@@ -106,10 +135,11 @@ describe('monitoring tasks api', () => {
     });
     expect(response.items.length).toBe(1);
     expect(response.items[0].externalId).toEqual(monitoringTaskExternalId);
-    expect(response.items[0].name).toEqual(monitoringTaskName);
+    expect(response.items[0].name).toEqual(monitoringTaskNameUpdated);
     expect(response.items[0].model).toEqual(
       expect.objectContaining(expectedResponseModel)
     );
+
     expect(response.items[0].interval).toEqual(testMtInterval);
     expect(response.items[0].overlap).toEqual(testMtOverlap);
   });