+++ packages/playground/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.4](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@7.0.3...@cognite/sdk-playground@7.0.4) (2023-11-28)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+## [7.0.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@7.0.2...@cognite/sdk-playground@7.0.3) (2023-11-24)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+## [7.0.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@7.0.1...@cognite/sdk-playground@7.0.2) (2023-11-23)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+## [7.0.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@7.0.0...@cognite/sdk-playground@7.0.1) (2023-11-21)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
 # [7.0.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@6.0.10...@cognite/sdk-playground@7.0.0) (2023-10-10)
 
 ### Features